// src/types/screens/moviesType.ts
export type PeliculasType = {
    creado: string
    director: string
    editado: string
    lanzamiento_fecha: string
    productora: string
    titulo: string
    url: string
    prologo: string
    episodio_id: number
    especies:string[]
    naves:string[]
    personajes:string[]
    planetas:string[]
    vehiculos:string[]
}

export type PeliculasInglesType = { 
    created: string
    director: string
    edited: string
    release_date: string
    producer: string
    title: string
    url: string
    opening_crawl: string
    episode_id: number
    species:string[]
    starships:string[]
    characters:string[]
    planets:string[]
    vehicles:string[]
}